requirements.txt
tensorflow-directml>=2.10.0
opencv-python>=4.5.0
mediapipe>=0.9.0
numpy>=1.21.0
scikit-learn>=1.0.0
tqdm>=4.62.0
Pillow>=8.3.0
keyboard>=0.13.5
matplotlib>=3.5.0
seaborn>=0.12.0
pandas>=1.3.0
joblib>=1.1.0
Flask>=2.2.0
